--- a/onediff_comfy_nodes/extras_nodes/nodes_nexfort_booster.py
+++ b/onediff_comfy_nodes/extras_nodes/nodes_nexfort_booster.py
@@ -7,7 +7,8 @@ compiler_modes = collections.OrderedDict(
     {
         "jit:disable-runtime-fusion:low-precision": "This compiles super quickly, but the performance might not be optimized very noticeably.",
         "jit:benchmark:low-precision:freezing:cudagraphs": "This compiles the model very quickly, but the performance might be not as good as `TorchInductor` optimized models.",
-        "max-autotune:low-precision": "This will deliver a good performance and adapt quickly to shape changes.",
+        "max-optimize:max-autotune:low-precision": "This will deliver a good performance and adapt quickly to shape changes.",
+        # "max-optimize:max-autotune:low-precision": "",
         "max-autotune:benchmark:low-precision:cudagraphs": "This is the most suggested combination of compiler modes. It will deliver a good balance between performance and compilation time.",
         "max-optimize:max-autotune:benchmark:low-precision:freezing:cudagraphs": "This is the most aggressive combination of compiler modes. It will deliver the best performance but might slow down the compilation significantly.",
     }