@lixiang007666 2024-06-13
- https://oneflow-static.oss-cn-beijing.aliyuncs.com/sd3/clip_g.safetensors
- https://oneflow-static.oss-cn-beijing.aliyuncs.com/sd3/clip_l.safetensors
- https://oneflow-static.oss-cn-beijing.aliyuncs.com/sd3/sd3_medium.safetensors
- https://oneflow-static.oss-cn-beijing.aliyuncs.com/sd3/t5xxl_fp8_e4m3fn.safetensors

@ccssu 这些是工作流里可能用到的模型。